requirements.txt
faiss_cpu
sentence_transformers
bitsandbytes
accelerate
pypdf
langchain==0.2.7
torch
transformers
dataclasses_json>=0.6.7,<0.7.0
Pillow>=9.0.0
torch
torchvision
gdown